refactor: move app shell components into app slices
