+++ src/app/ui/AppActivityBar/ActivityBar.tsx
