+++ src/app/ui/AppStatusBar/AppStatusBar.tsx
@@ -4,8 +4,8 @@
  */
 
 import { useAtomValue } from 'jotai';
-import { StatusBar } from '@/components/ide/StatusBar.tsx';
 import { parseProgressAtom } from '@/entities/AppView/model/atoms';
+import { StatusBar } from './StatusBar';
 
 export function AppStatusBar() {
   const parseProgress = useAtomValue(parseProgressAtom);
